Move view related plans
