+++ src/query/planner/src/plans/create_view.rs
@@ -11,22 +11,22 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
 use std::sync::Arc;
 
 use common_datavalues::DataSchema;
 use common_datavalues::DataSchemaRef;
 
 #[derive(serde::Serialize, serde::Deserialize, Clone, Debug, PartialEq, Eq)]
-pub struct DropViewPlan {
-    pub if_exists: bool,
+pub struct CreateViewPlan {
+    pub if_not_exists: bool,
     pub tenant: String,
     pub catalog: String,
     pub database: String,
     pub viewname: String,
+    pub subquery: String,
 }
 
-impl DropViewPlan {
+impl CreateViewPlan {
     pub fn schema(&self) -> DataSchemaRef {
         Arc::new(DataSchema::empty())
     }
