+++ src/query/planner/src/plans/mod.rs
@@ -15,12 +15,14 @@
 mod alter_table_cluster_key;
 mod alter_udf;
 mod alter_user;
+mod alter_view;
 mod call;
 mod create_database;
 mod create_role;
 mod create_stage;
 mod create_udf;
 mod create_user;
+mod create_view;
 mod describe_table;
 mod drop_database;
 mod drop_role;
@@ -29,6 +31,7 @@ mod drop_table;
 mod drop_table_cluster_key;
 mod drop_udf;
 mod drop_user;
+mod drop_view;
 mod exists_table;
 mod grant_privilege;
 mod grant_role;
@@ -51,12 +54,14 @@ mod use_database;
 pub use alter_table_cluster_key::AlterTableClusterKeyPlan;
 pub use alter_udf::AlterUDFPlan;
 pub use alter_user::AlterUserPlan;
+pub use alter_view::AlterViewPlan;
 pub use call::CallPlan;
 pub use create_database::CreateDatabasePlan;
 pub use create_role::CreateRolePlan;
 pub use create_stage::CreateStagePlan;
 pub use create_udf::CreateUDFPlan;
 pub use create_user::CreateUserPlan;
+pub use create_view::CreateViewPlan;
 pub use describe_table::DescribeTablePlan;
 pub use drop_database::DropDatabasePlan;
 pub use drop_role::DropRolePlan;
@@ -65,6 +70,7 @@ pub use drop_table::DropTablePlan;
 pub use drop_table_cluster_key::DropTableClusterKeyPlan;
 pub use drop_udf::DropUDFPlan;
 pub use drop_user::DropUserPlan;
+pub use drop_view::DropViewPlan;
 pub use exists_table::ExistsTablePlan;
 pub use grant_privilege::GrantPrivilegePlan;
 pub use grant_role::GrantRolePlan;
