+++ src/query/service/src/interpreters/interpreter_view_alter.rs
@@ -17,11 +17,11 @@ use std::sync::Arc;
 
 use common_exception::ErrorCode;
 use common_exception::Result;
-use common_legacy_planners::AlterViewPlan;
 use common_meta_app::schema::CreateTableReq;
 use common_meta_app::schema::DropTableReq;
 use common_meta_app::schema::TableMeta;
 use common_meta_app::schema::TableNameIdent;
+use common_planner::plans::AlterViewPlan;
 
 use crate::interpreters::Interpreter;
 use crate::pipelines::PipelineBuildResult;
