+++ src/query/service/src/interpreters/interpreter_view_drop.rs
@@ -16,9 +16,9 @@ use std::sync::Arc;
 
 use common_exception::ErrorCode;
 use common_exception::Result;
-use common_legacy_planners::DropViewPlan;
 use common_meta_app::schema::DropTableReq;
 use common_meta_app::schema::TableNameIdent;
+use common_planner::plans::DropViewPlan;
 
 use crate::interpreters::Interpreter;
 use crate::pipelines::PipelineBuildResult;
