+++ src/query/service/src/sql/planner/binder/ddl/view.rs
@@ -16,9 +16,9 @@ use common_ast::ast::AlterViewStmt;
 use common_ast::ast::CreateViewStmt;
 use common_ast::ast::DropViewStmt;
 use common_exception::Result;
-use common_legacy_planners::AlterViewPlan;
-use common_legacy_planners::CreateViewPlan;
-use common_legacy_planners::DropViewPlan;
+use common_planner::plans::AlterViewPlan;
+use common_planner::plans::CreateViewPlan;
+use common_planner::plans::DropViewPlan;
 
 use crate::sql::binder::Binder;
 use crate::sql::planner::semantic::normalize_identifier;
